% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{as_matrix}
\alias{as_matrix}
\title{Coerce a gridded object to a matrix}
\usage{
as_matrix(x, ...)
}
\arguments{
\item{x}{A gridded object. Currently methods exist for classes data.frame,
Raster, and stars.}

\item{...}{Additional arguments to pass to coercion function.}
}
\value{
A matrix.
}
\description{
Coerce a gridded object to a matrix
}
