% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calc_climatology}
\alias{calc_climatology}
\title{Time-averaged footprint}
\usage{
calc_climatology(x, grid, weights = NULL)
}
\arguments{
\item{x}{A list of footprint matrices.}

\item{grid}{A list of length two containing matrices of equal dimensions,
indicating x and y coordinates. Template returned by \link{grid_init}.}

\item{weights}{A vector of same length as x}
}
\description{
Time-averaged footprint
}
