% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calc_footprint}
\alias{calc_footprint}
\title{Calculate two-dimensional footprints}
\usage{
calc_footprint(
  grid,
  wd,
  ustar,
  mo_length,
  v_sigma,
  z,
  zd,
  zo,
  ws = NULL,
  blh = NULL,
  model = c("H00", "K15", "KM01")
)
}
\arguments{
\item{grid}{A list of length two containing matrices of equal dimensions,
indicating x and y coordinates. Template returned by \link{grid_init}.}

\item{wd}{A numeric vector, wind direction in degrees from North.}

\item{ustar}{A numeric vector, friction velocity in ms+1s-1.}

\item{mo_length}{A numeric vector, Monin-Obukhov length in m.}

\item{v_sigma}{A numeric vector, standard deviation of the crosswind
component.}

\item{z}{A numeric value, tower height in m.}

\item{zd}{A numeric value, zero-plane displacement height in m.}

\item{zo}{The aerodynamic roughness length in m.}

\item{ws}{A numeric vector, wind speed in ms+1s-1.}

\item{blh}{A numeric value, boundary layer height in m.}

\item{model}{A character string naming the model to be used in footprint
calculations. Can be "KM01" for the Kormann & Meixner (2001) model or "H00"
for the Hsieh et al. (2000) model.}
}
\description{
Create a two-dimensional matrix of footprint probabilities.
}
\details{
Support only for square grids at the moment.
}
