% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calc_footprint_hsieh}
\alias{calc_footprint_hsieh}
\title{Two-dimensional footprint calculation}
\usage{
calc_footprint_hsieh(grid, wd, ustar, mo_length, v_sigma, zm, zd, zo, ...)
}
\arguments{
\item{grid}{A list of matrices as returned by \code{\link[=grid_init]{grid_init()}}}

\item{wd}{A numeric vector, wind direction in degrees from North}

\item{ustar}{A numeric vector, friction velocity in ms+1s-1}

\item{mo_length}{A numeric vector, Monin-Obukhov length in m}

\item{v_sigma}{A numeric vector, standard deviation of the crosswind
component}

\item{zm}{A numeric vector, measurement height in m}

\item{zd}{A numeric vector, zero-plane displacement height in m}

\item{zo}{A numeric vector, aerodynamic roughness length in m}

\item{...}{To be left empty, for flexible calls to a footprint function}
}
\value{
A matrix with same dimensions as each item of grid
}
\description{
Two-dimensional footprint calculation
}
\references{
Detto, M., Montaldo, N., Albertson, J. D., Mancini, M., & Katul, G. (2006).
Soil moisture and vegetation controls on evapotranspiration in a
heterogeneous Mediterranean ecosystem on Sardinia, Italy. Water Resources
Research, 42(8), 1–16. https://doi.org/10.1029/2005WR004693

Hsieh, C.-I., Katul, G., & Chi, T. (2000). An approximate analytical model
for footprint estimation of scalar fluxes in thermally stratified atmospheric
flows. Advances in Water Resources, 23(7), 765–772.
https://doi.org/10.1016/S0309-1708(99)00042-1
}
\seealso{
Other footprints: 
\code{\link{calc_footprint_kljun}()},
\code{\link{calc_footprint_kormann}()}
}
\concept{footprints}
