% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calc_footprint_kljun}
\alias{calc_footprint_kljun}
\title{Two-dimensional footprint calculation}
\usage{
calc_footprint_kljun(grid, wd, ustar, mo_length, v_sigma, blh, zm, zd, zo, ...)
}
\arguments{
\item{grid}{A list of matrices as returned by \code{\link[=grid_init]{grid_init()}}}

\item{wd}{A numeric vector, wind direction in degrees from North}

\item{ustar}{A numeric vector, friction velocity in ms+1s-1}

\item{mo_length}{A numeric vector, Monin-Obukhov length in m}

\item{v_sigma}{A numeric vector, standard deviation of the crosswind
component}

\item{blh}{A numeric vector, boundary layer height in m}

\item{zm}{A numeric vector, measurement height in m}

\item{zd}{A numeric vector, zero-plane displacement height in m}

\item{zo}{A numeric vector, aerodynamic roughness length in m}

\item{...}{To be left empty, for flexible calls to a footprint function}
}
\value{
A matrix with same dimensions as each item of grid
}
\description{
Two-dimensional footprint calculation
}
\details{
The authors assert the following criteria must be met for their model to be
valid:
\itemize{
\item ustar > 0.1
\item (zm - zd) / mo_length >= -15.5
}
}
\references{
Kljun, N., Calanca, P., Rotach, M. W., & Schmid, H. P. (2015). A simple
two-dimensional parameterisation for Flux Footprint Prediction (FFP).
Geoscientific Model Development, 8(11),
3695–3713. https://doi.org/10.5194/gmd-8-3695-2015
}
\seealso{
Other footprints: 
\code{\link{calc_footprint_hsieh}()},
\code{\link{calc_footprint_kormann}()}
}
\concept{footprints}
