% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calc_footprint_kormann}
\alias{calc_footprint_kormann}
\title{Two-dimensional footprint calculation}
\usage{
calc_footprint_kormann(grid, wd, ws, ustar, mo_length, v_sigma, zm, zd, ...)
}
\arguments{
\item{grid}{A list of matrices as returned by \code{\link[=grid_init]{grid_init()}}}

\item{wd}{A numeric vector, wind direction in degrees from North}

\item{ws}{A numeric vector, wind speed in m+1s-1}

\item{ustar}{A numeric vector, friction velocity in m+1s-1}

\item{mo_length}{A numeric vector, Monin-Obukhov length in m}

\item{v_sigma}{A numeric vector, standard deviation of the crosswind
component}

\item{zm}{A numeric vector, measurement height in m}

\item{zd}{A numeric vector, zero-plane displacement height in m}

\item{...}{To be left empty, for flexible calls to a footprint function}
}
\value{
A matrix with same dimensions as each item of grid
}
\description{
Two-dimensional footprint calculation
}
\references{
Kormann, R., & Meixner, F. X. (2001). An Analytical Footprint Model For
Non-Neutral Stratification. Boundary-Layer Meteorology, 99(2), 207–224.
\url{https://doi.org/10.1023/A:1018991015119}
}
\seealso{
Other footprints: 
\code{\link{calc_footprint_hsieh}()},
\code{\link{calc_footprint_kljun}()}
}
\concept{footprints}
