% Generated by roxygen2 (4.0.1): do not edit by hand
\name{citation_location}
\alias{citation_location}
\title{Find location of particular citation in text}
\usage{
citation_location(full_text, key, breaks = c("Introduction", "Methods",
  "Results", "Discussion"))
}
\arguments{
\item{full_text}{This is an array of text that is usually grabbed
from \code{grab_text()} function.}

\item{key}{The sequence you are searching for in the \code{full_text}}

\item{breaks}{The sequence of section headers to partition the paper into.  By default,
the text is broken into "Introduction", "Methods", "Results", and "Discussion"}
}
\value{
A list with the number of times the \code{key} was found in each section (defined
 by \code{breaks}) in \code{full_text}.
}
\description{
This is a handy function to find out where a citation is
 located in a manuscript.
}
\note{
This function will allow you to specify the partitions of the manuscript in which
 to look for a particular citation (or phrase).  It should be noted that you need to
 try alternate citation types (e.g., Smith (1997), (Smith 1997), Smith .... (1997)), this
 function isn't that cleaver...
}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}

