% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{footprint-package}
\alias{footprint-package}
\title{Package for the analysis of intellectual space}
\description{
This package provides a set of routines necessary to quantify
 intellectual space based upon text mining manuscript abstracts.
}
\details{
\tabular{ll}{
Package: \tab footprint\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-05-29\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}

There are some very good examples of the components of this package are used
 in the vignettes for this package.
}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}
\keyword{package}

