% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fp_rasterize}
\alias{fp_rasterize}
\title{Rasterize a footprint matrix}
\usage{
fp_rasterize(x, grid, coords, crs)
}
\arguments{
\item{x}{A footprint object}

\item{grid}{A list of length two containing matrices of equal dimensions,
indicating x and y coordinates. Template returned by \link{fp_grid}.}

\item{coords}{A numeric vector of length two}

\item{crs}{A character or object of class CRS.}
}
\description{
Rasterize a footprint matrix
}
