% Generated by roxygen2 (4.0.1): do not edit by hand
\name{grab_text}
\alias{grab_text}
\title{Grab raw text from pdf}
\usage{
grab_text(pdf, pdf2txt = "pdf2txt.py", remove_nulls = TRUE,
  remove_singles = TRUE, ...)
}
\arguments{
\item{pdf}{The path to the pdf file}

\item{pdf2txt}{The shell command to run the python script pdf2txt.py (default)}

\item{remove_nulls}{A flag indicating that empty lines should be ignored (default = TRUE)}

\item{remove_singles}{A flag indicating that lines with only one character on them should be
ignored (default=TRUE)}

\item{...}{Ignored}
}
\value{
The raw text from the file.
}
\description{
This function returns the entire text from a pdf.
}
\note{
This function is the main function through which pdf text is to be filtered.  This funciton
 requires pdf2txt.py to be installed on the machine.
}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}

