% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make_clean_tdm}
\alias{make_clean_tdm}
\title{Creates a clean term document matrix}
\usage{
make_clean_tdm(text)
}
\arguments{
\item{text}{A vector of character values for raw text.}
}
\value{
A cleaned up Term Document Matrix
}
\description{
This function seeks to create a cleaned up
term document matrix from raw text.
}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}

