% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make_conceptual_tdm}
\alias{make_conceptual_tdm}
\title{Makes a term document matrix out of a folder full of PDF's}
\usage{
make_conceptual_tdm(path, ...)
}
\arguments{
\item{path}{The path to the folder.  The PDF's must be labelled *.pdf.}

\item{...}{Other parameters to be passed to \code{grab_text()} and \code{grab_chunk()}
functions.  See their descriptions to what they allow to be passed or what is
used as defaults.}
}
\value{
A term document matrix.
}
\description{
This is more of a convienence function but allows you to quickly
make a term document matrix out of a folder of PDF's
}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}

