% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mask_source_area}
\alias{mask_source_area}
\title{Mask footprint to its analytical source area}
\usage{
mask_source_area(x, p = 0.85, mask_value = NA)
}
\arguments{
\item{x}{The footprint as a numeric matrix or RasterLayer.}

\item{p}{A numeric value between 0 and 1, recommended by Kljun et al. (2015)
to be less than 0.9. Defaults to 0.85.}

\item{mask_value}{A value to assign cells beyond the maximum percent weight.
Defaults to NA, but 0 may be useful in some situations.}
}
\value{
An object of the same class as x, either a numeric matrix or
RasterLayer.
}
\description{
As the analytical footprint approaches 100\%, the coverage increases rapidly
over area that contributes little to the total measured flux (Matthes et al.,
2014). Thus, it may be desirable to consider a percantage of the calculated
footprint as the analytical footprint. This is acheived by masking the
footprint to a given maximum cumulative weight.
}
