% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_tidy_matrix}
\alias{plot_tidy_matrix}
\title{Plot a generic matrix using ggplot}
\usage{
plot_tidy_matrix(x, trans = NA)
}
\arguments{
\item{x}{A matrix.}

\item{trans}{Name of a transformation object as defined in
ggplot2::continuous_scale. "sqrt" and "log" are particularly useful.}
}
\value{
A ggplot object.
}
\description{
Plot a generic matrix using ggplot
}
