% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_grid}
\alias{read_grid}
\title{Read grid coordinates from two .txt files}
\usage{
read_grid(path, names = c("x", "y"))
}
\arguments{
\item{path}{A character vector containing a single path}

\item{names}{A character vector of length two, the file names of grid
elements (omitting the extension)}
}
\description{
Read grid coordinates from two .txt files
}
