% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_matrix}
\alias{read_matrix}
\title{Read whitespace-separated columns into a matrix}
\usage{
read_matrix(file, trunc = 0)
}
\arguments{
\item{file}{A path to a file.}

\item{trunc}{Number of decimal places used to integerize data contents. If
contents were not truncated, set to 0 (the default). A common value is 9.}
}
\description{
Read whitespace-separated columns into a matrix
}
