% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{snap_to_grid}
\alias{snap_to_grid}
\title{Resample a raster to a standard matrix grid}
\usage{
snap_to_grid(x, grid, coords, method = "ngb")
}
\arguments{
\item{x}{A Raster object to be resampled.}

\item{grid}{A list of matrices as returned by \code{\link[=grid_init]{grid_init()}}.}

\item{coords}{A numeric vector of length two indicating the spatial
coordinates of the "center" grid cell.}

\item{method}{Name of the resampling method passed to raster::resample. Can
be "ngb" (the default) or "bilinear".}
}
\value{
A matrix that matches spatial dimensions of grid.
}
\description{
Resample a raster to a standard matrix grid
}
