% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{source_area_polygon}
\alias{source_area_polygon}
\title{Convert footprint to a polygon representing a given source area}
\usage{
source_area_polygon(x, p = 0.85)
}
\arguments{
\item{x}{The footprint as a numeric matrix or RasterLayer.}

\item{p}{A numeric value between 0 and 1, recommended by Kljun et al. (2015)
to be less than 0.9. Defaults to 0.85.}
}
\description{
Convert footprint to a polygon representing a given source area
}
