% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{summarize_cover}
\alias{summarize_cover}
\title{Footprint-weighted land cover}
\usage{
summarize_cover(x, y, type = c("factor", "numeric"), levels, ignore_levels = 0)
}
\arguments{
\item{x}{A numeric matrix}

\item{y}{A numeric matrix}

\item{type}{The type of convolution to take place, either "factor" or
"numeric"}

\item{levels}{A vector}

\item{ignore_levels}{A vector}
}
\description{
Footprint-weighted land cover
}
