% Generated by roxygen2 (4.0.1): do not edit by hand
\name{term_frequency}
\alias{term_frequency}
\title{Get word frequency from chunk of text}
\usage{
term_frequency(text, show_cloud = FALSE)
}
\arguments{
\item{text}{The chunk of text to use}

\item{show_cloud}{An optional term to plot as a wordcloud if necessary}
}
\value{
A \code{data.frame} of word frequencies in the chunk
}
\description{
This takes a set of text items and returns a \code{data.frame} from
 it with important words in reverse importance of frequency.
}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}

