% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{write_matrix}
\alias{write_matrix}
\title{Write a matrix to a delimited file}
\usage{
write_matrix(x, path, trunc = 0, compress = TRUE)
}
\arguments{
\item{x}{A matrix object.}

\item{path}{A path to write to.}

\item{trunc}{Number of decimal places used to integerize data contents. If
set to 0 (the default), no changes are made. A common value is 9.}

\item{compress}{Use gzip compression? Must be either TRUE or FALSE. If TRUE
(recommended), a ".gz" extension will be added to the end of file.}
}
\description{
Write a matrix to a delimited file
}
